requirements.txt
Appium-Python-Client==0.38
atomicwrites==1.3.0
attrs==18.2.0
Logbook==1.4.3
more-itertools==6.0.0
pluggy==0.9.0
ptest==1.9.4
py==1.8.0
pytest==4.3.0
PyYAML==3.13
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
